refactor(blockchain): move ChainEventSnapshot into events module
Colocate the pre-call snapshot type (and HEAD_EVENT_RECENCY_SLOTS) with
the ChainEvent/EventBus definitions it depends on, keeping all chain-event
machinery in events.rs. The actor references it via crate::events; the
snapshot's diff/gate tests move alongside it. No behavior change.
