+++ crates/blockchain/src/events.rs
@@ -9,9 +9,13 @@
 //! a slow subscriber loses events (the bounded broadcast channel overwrites
 //! its backlog) rather than back-pressuring consensus.
 
+use ethlambda_storage::Store;
+use ethlambda_types::ShortRoot;
+use ethlambda_types::checkpoint::Checkpoint;
 use ethlambda_types::primitives::H256;
 use serde::Serialize;
 use tokio::sync::broadcast;
+use tracing::warn;
 
 /// Wire-visible topic names for chain events.
 ///
@@ -125,9 +129,111 @@ impl Default for EventBus {
     }
 }
 
+/// Suppresses `head` events whose slot has fallen too far behind the wall
+/// clock: during startup catch-up or backfill, fork choice can walk through
+/// many historical heads on its way to the tip, and none of those are
+/// interesting to a live subscriber. Consumers that need to track sync
+/// progress should watch `block` events instead, which are never gated on
+/// recency. Mirrors Lighthouse's recency filter on its head SSE event
+/// (`EARLY_ATTESTER_CACHE_HISTORIC_SLOTS`), but with a wider window since a
+/// lagging head is far more common here during multi-slot catch-up ticks.
+const HEAD_EVENT_RECENCY_SLOTS: u64 = 32;
+
+/// Pre-call snapshot of the store values the chain-event bus reports on.
+///
+/// The actor — not the store — publishes chain events: it captures this
+/// snapshot before a store call (`store::on_tick`, `store::on_block`) and
+/// diffs the store against it afterwards, so `store.rs` needs no event
+/// plumbing. Consequences of diffing at this level, both intentional:
+///
+/// - Multiple head moves within one store call coalesce into a single `head`
+///   event; subscribers only care about the latest.
+/// - The proposer's pre-build catch-up (`get_proposal_head`) runs outside any
+///   snapshot window, so a head move never surfaces before the block it
+///   belongs to exists; the block's own import then emits the final head.
+pub(crate) struct ChainEventSnapshot {
+    head: H256,
+    justified: Checkpoint,
+    finalized: Checkpoint,
+}
+
+impl ChainEventSnapshot {
+    pub(crate) fn capture(store: &Store) -> Self {
+        Self {
+            head: store.head().expect("head block exists"),
+            justified: store
+                .latest_justified()
+                .expect("latest justified checkpoint exists"),
+            finalized: store
+                .latest_finalized()
+                .expect("latest finalized checkpoint exists"),
+        }
+    }
+
+    /// Emit one event per value that changed since the snapshot, in a fixed
+    /// order: `head` → `justified_checkpoint` → `finalized_checkpoint`.
+    /// (`block` is emitted separately by the import path, ahead of this diff.)
+    ///
+    /// `wall_clock_slot` is the caller's current slot, used only to gate the
+    /// `head` event against [`HEAD_EVENT_RECENCY_SLOTS`]; the other events are
+    /// ungated.
+    pub(crate) fn diff_and_emit(&self, store: &Store, events: &EventBus, wall_clock_slot: u64) {
+        let head = store.head().expect("head block exists");
+        if head != self.head {
+            // Read the header once and reuse it for slot and parent_root so
+            // they stay consistent.
+            if let Some(header) = store
+                .get_block_header(&head)
+                .expect("block header read should succeed")
+            {
+                // Skip stale heads (catch-up/backfill): see HEAD_EVENT_RECENCY_SLOTS.
+                if header.slot + HEAD_EVENT_RECENCY_SLOTS >= wall_clock_slot {
+                    events.emit(ChainEvent::Head {
+                        slot: header.slot,
+                        root: head,
+                        parent_root: header.parent_root,
+                    });
+                }
+            } else {
+                warn!(
+                    head_root = %ShortRoot(&head.0),
+                    "Head header missing while emitting head event; skipping"
+                );
+            }
+        }
+
+        let justified = store
+            .latest_justified()
+            .expect("latest justified checkpoint exists");
+        if justified != self.justified {
+            events.emit(ChainEvent::JustifiedCheckpoint {
+                slot: justified.slot,
+                root: justified.root,
+            });
+        }
+
+        let finalized = store
+            .latest_finalized()
+            .expect("latest finalized checkpoint exists");
+        if finalized != self.finalized {
+            events.emit(ChainEvent::FinalizedCheckpoint {
+                slot: finalized.slot,
+                root: finalized.root,
+            });
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
+    use ethlambda_storage::{ForkCheckpoints, backend::InMemoryBackend};
+    use ethlambda_types::{
+        block::{Block, BlockBody, MultiMessageAggregate, SignedBlock},
+        state::State,
+    };
+    use std::sync::Arc;
+    use tokio::sync::broadcast::error::TryRecvError;
 
     fn head_event(slot: u64) -> ChainEvent {
         ChainEvent::Head {
@@ -200,5 +306,201 @@ mod tests {
         assert!(json["parent_root"].is_string());
         assert!(json.get("event").is_none());
         assert!(json.get("data").is_none());
+    }
+
+    fn test_store() -> Store {
+        let genesis_state = State::from_genesis(1000, vec![]);
+        Store::from_anchor_state(Arc::new(InMemoryBackend::new()), genesis_state)
+    }
+
+    /// Insert a header-only block at `root` so head-header reads resolve.
+    fn insert_test_block(store: &mut Store, root: H256, slot: u64, parent_root: H256) {
+        let signed_block = SignedBlock {
+            message: Block {
+                slot,
+                proposer_index: 0,
+                parent_root,
+                state_root: H256::ZERO,
+                body: BlockBody::default(),
+            },
+            proof: MultiMessageAggregate::default(),
+        };
+        store
+            .insert_signed_block(root, signed_block)
+            .expect("insert test block should succeed");
+    }
+
+    #[test]
+    fn chain_event_diff_emits_nothing_when_unchanged() {
+        let store = test_store();
+        let bus = EventBus::new(8);
+        let mut rx = bus.subscribe();
+
+        let snapshot = ChainEventSnapshot::capture(&store);
+        snapshot.diff_and_emit(&store, &bus, 0);
+
+        assert!(matches!(rx.try_recv(), Err(TryRecvError::Empty)));
+    }
+
+    /// A head whose header cannot be read is skipped (warn), while checkpoint
+    /// moves still emit.
+    #[test]
+    fn chain_event_diff_skips_head_with_missing_header() {
+        let mut store = test_store();
+        let bus = EventBus::new(8);
+        let mut rx = bus.subscribe();
+
+        let snapshot = ChainEventSnapshot::capture(&store);
+
+        // Point the head at a root with no stored header; advance finalized to
+        // a real block so its event still fires.
+        let orphan_head = H256([7u8; 32]);
+        let genesis = store.head().expect("store head exists");
+        let finalized_root = H256([8u8; 32]);
+        insert_test_block(&mut store, finalized_root, 1, genesis);
+        let finalized = Checkpoint {
+            root: finalized_root,
+            slot: 1,
+        };
+        store
+            .update_checkpoints(ForkCheckpoints::new(orphan_head, None, Some(finalized)))
+            .expect("update_checkpoints should succeed");
+
+        snapshot.diff_and_emit(&store, &bus, 1);
+
+        match rx.try_recv().unwrap() {
+            ChainEvent::FinalizedCheckpoint { slot, root } => {
+                assert_eq!((slot, root), (1, finalized_root));
+            }
+            other => panic!("expected finalized_checkpoint only, got: {other:?}"),
+        }
+        assert!(matches!(rx.try_recv(), Err(TryRecvError::Empty)));
+    }
+
+    /// A head far below the wall-clock slot (catch-up/backfill) emits no
+    /// `head` event, but `justified_checkpoint`/`finalized_checkpoint` still
+    /// fire since only `head` is gated.
+    #[test]
+    fn chain_event_diff_gates_stale_head() {
+        let mut store = test_store();
+        let genesis = store.head().expect("store head exists");
+        let bus = EventBus::new(8);
+        let mut rx = bus.subscribe();
+
+        let snapshot = ChainEventSnapshot::capture(&store);
+
+        let new_root = H256([9u8; 32]);
+        insert_test_block(&mut store, new_root, 1, genesis);
+        let checkpoint = Checkpoint {
+            root: new_root,
+            slot: 1,
+        };
+        store
+            .update_checkpoints(ForkCheckpoints::new(
+                new_root,
+                Some(checkpoint),
+                Some(checkpoint),
+            ))
+            .expect("update_checkpoints should succeed");
+
+        // Wall clock far ahead of the new head's slot (1): well past
+        // HEAD_EVENT_RECENCY_SLOTS, so the head event must be suppressed.
+        let wall_clock_slot = 1 + HEAD_EVENT_RECENCY_SLOTS + 100;
+        snapshot.diff_and_emit(&store, &bus, wall_clock_slot);
+
+        match rx.try_recv().unwrap() {
+            ChainEvent::JustifiedCheckpoint { slot, root } => {
+                assert_eq!((slot, root), (1, new_root));
+            }
+            other => panic!("expected justified_checkpoint first, got: {other:?}"),
+        }
+        match rx.try_recv().unwrap() {
+            ChainEvent::FinalizedCheckpoint { slot, root } => {
+                assert_eq!((slot, root), (1, new_root));
+            }
+            other => panic!("expected finalized_checkpoint second (head gated), got: {other:?}"),
+        }
+        assert!(matches!(rx.try_recv(), Err(TryRecvError::Empty)));
+    }
+
+    /// A head within the recency window emits normally, stated explicitly
+    /// against the gate for clarity.
+    #[test]
+    fn chain_event_diff_emits_recent_head() {
+        let mut store = test_store();
+        let genesis = store.head().expect("store head exists");
+        let bus = EventBus::new(8);
+        let mut rx = bus.subscribe();
+
+        let snapshot = ChainEventSnapshot::capture(&store);
+
+        let new_root = H256([9u8; 32]);
+        insert_test_block(&mut store, new_root, 1, genesis);
+        store
+            .update_checkpoints(ForkCheckpoints::head_only(new_root))
+            .expect("update_checkpoints should succeed");
+
+        // Wall clock equal to the head's own slot: as recent as it gets.
+        snapshot.diff_and_emit(&store, &bus, 1);
+
+        match rx.try_recv().unwrap() {
+            ChainEvent::Head { slot, root, .. } => {
+                assert_eq!((slot, root), (1, new_root));
+            }
+            other => panic!("expected head event, got: {other:?}"),
+        }
+        assert!(matches!(rx.try_recv(), Err(TryRecvError::Empty)));
+    }
+
+    /// Boundary case: `new_head_slot + HEAD_EVENT_RECENCY_SLOTS >=
+    /// wall_clock_slot` is the emit condition, so equality still emits, while
+    /// one slot further pushes it over into "gated".
+    #[test]
+    fn chain_event_diff_head_recency_boundary() {
+        let head_slot = 1;
+
+        // Exactly at the boundary: emits (`>=` includes equality).
+        {
+            let mut store = test_store();
+            let genesis = store.head().expect("store head exists");
+            let bus = EventBus::new(8);
+            let mut rx = bus.subscribe();
+            let snapshot = ChainEventSnapshot::capture(&store);
+
+            let new_root = H256([9u8; 32]);
+            insert_test_block(&mut store, new_root, head_slot, genesis);
+            store
+                .update_checkpoints(ForkCheckpoints::head_only(new_root))
+                .expect("update_checkpoints should succeed");
+
+            let wall_clock_slot = head_slot + HEAD_EVENT_RECENCY_SLOTS;
+            snapshot.diff_and_emit(&store, &bus, wall_clock_slot);
+
+            match rx.try_recv().unwrap() {
+                ChainEvent::Head { slot, .. } => assert_eq!(slot, head_slot),
+                other => panic!("expected head event at the boundary, got: {other:?}"),
+            }
+            assert!(matches!(rx.try_recv(), Err(TryRecvError::Empty)));
+        }
+
+        // One slot past the boundary: gated.
+        {
+            let mut store = test_store();
+            let genesis = store.head().expect("store head exists");
+            let bus = EventBus::new(8);
+            let mut rx = bus.subscribe();
+            let snapshot = ChainEventSnapshot::capture(&store);
+
+            let new_root = H256([9u8; 32]);
+            insert_test_block(&mut store, new_root, head_slot, genesis);
+            store
+                .update_checkpoints(ForkCheckpoints::head_only(new_root))
+                .expect("update_checkpoints should succeed");
+
+            let wall_clock_slot = head_slot + HEAD_EVENT_RECENCY_SLOTS + 1;
+            snapshot.diff_and_emit(&store, &bus, wall_clock_slot);
+
+            assert!(matches!(rx.try_recv(), Err(TryRecvError::Empty)));
+        }
     }
 }
